requirements.txt
Django
black
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg[binary,pool]
gunicorn
reportlab
matplotlib
Pillow